refactor: split oversized files + fix orphan screens (metromap cleanup)
- settings/specialist.tsx: 751L -> 420L (extract AvatarUploader, ContactMethodsList, NotificationPreferences)
- requests/new.tsx: 604L -> 283L (extract CityFnsServicePicker, FileUploadSection)
- requests/[id]/detail.tsx: 554L -> 371L (extract ThreadsList, SpecialistRecommendations)
- listing/[id].tsx: removed (dead code, zero incoming router.push calls, placeholder content)
- /notifications: already linked from Header.tsx (mobile + desktop) + HeaderHome.tsx

Closes #1198 #1199 #1200 #1201
